backend/requirements.txt
Django>=4.2.16
psycopg2==2.9.7
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.0
djangorestframework==3.15.2
#environs
